refactor: Simplify DE (closes #690)
